Create db access utils, move generic code from explorer to utils
